pyproject.toml
[tool.poetry]
name = "apiproject"
version = "0.1.0"
description = "Projeto para o processo de seleçao da Logap"
authors = ["Andre Bezerra <amelco.herman@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.4"
djangorestframework = "^3.12.4"
psycopg2-binary = "^2.8.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
